Refactored `waitReadyToProcessRecord()` and the call chain down to right before `waitUntilValueSchemaAvailable()` from `StoreIngestionTask` into `StorePartitionDataReceiver`. 🧹
